+++ snooty/types.py
@@ -1,24 +1,9 @@
-import enum
 import hashlib
 import logging
 import re
-from docutils.nodes import make_id
-from collections import defaultdict
 from dataclasses import dataclass, field
-from pathlib import Path, PurePath, PurePosixPath
-from typing import (
-    Dict,
-    DefaultDict,
-    NamedTuple,
-    Set,
-    List,
-    Iterator,
-    Sequence,
-    MutableSequence,
-    Tuple,
-    Optional,
-    Match,
-)
+from pathlib import Path, PurePosixPath
+from typing import Dict, List, MutableSequence, Tuple, Optional, Match
 from .diagnostics import (
     Diagnostic,
     UnmarshallingError,
@@ -26,10 +11,9 @@ from .diagnostics import (
     ConstantNotDeclared,
 )
 from typing_extensions import Protocol
-import urllib.parse
 import toml
 from .flutter import checked, check_type, LoadError
-from . import n, intersphinx
+from . import n
 from .n import SerializableType as ST
 
 SerializableType = ST
@@ -38,10 +22,6 @@ PAT_GIT_MARKER = re.compile(r"^<<<<<<< .*?^=======\n.*?^>>>>>>>", re.M | re.S)
 PAT_FILE_EXTENSIONS = re.compile(r"\.((txt)|(rst)|(yaml))$")
 BuildIdentifierSet = Dict[str, Optional[str]]
 logger = logging.getLogger(__name__)
-
-#: Indicates the target protocol of a target: either a file local to the
-#: current project, or a URL (from an intersphinx inventory).
-TargetType = enum.Enum("TargetType", ("fileid", "url"))
 
 
 class EmbeddedRstParser(Protocol):
@@ -117,262 +97,12 @@ class StaticAsset:
             self._checksum = hashlib.blake2b(self._data, digest_size=32).hexdigest()
 
 
-@dataclass
-class TargetDatabase:
-    """A database of targets known to this project."""
-
-    class Result(NamedTuple):
-        type: TargetType
-        result: str
-        canonical_target_name: str
-        title: Sequence[n.InlineNode]
-
-    class LocalDefinition(NamedTuple):
-        canonical_name: str
-        fileid: FileId
-        title: Sequence[n.InlineNode]
-
-    intersphinx_inventories: Dict[str, intersphinx.Inventory] = field(
-        default_factory=dict
-    )
-    local_definitions: DefaultDict[str, List[LocalDefinition]] = field(
-        default_factory=lambda: defaultdict(list)
-    )
-
-    def __contains__(self, key: str) -> bool:
-        key = normalize_target(key)
-        if key in self.local_definitions:
-            return True
-
-        for inventory in self.intersphinx_inventories.values():
-            if key in inventory:
-                return True
-
-        return False
-
-    def __getitem__(self, key: str) -> Sequence["TargetDatabase.Result"]:
-        key = normalize_target(key)
-        results: List[TargetDatabase.Result] = []
-
-        # Check to see if the target is defined locally
-        try:
-            results.extend(
-                TargetDatabase.Result(
-                    TargetType.fileid,
-                    fileid.without_known_suffix,
-                    canonical_target_name,
-                    title,
-                )
-                for canonical_target_name, fileid, title in self.local_definitions[key]
-            )
-        except KeyError:
-            pass
-
-        # Get URL from intersphinx inventories
-        for inventory in self.intersphinx_inventories.values():
-            if key in inventory:
-                base_url = inventory.base_url
-                entry = inventory[key]
-                url = urllib.parse.urljoin(base_url, entry.uri)
-                title: List[n.InlineNode] = [n.Text((-1,), entry.display_name)]
-                results.append(
-                    TargetDatabase.Result(TargetType.url, url, entry.name, title)
-                )
-
-        return results
-
-    def define_local_target(
-        self,
-        domain: str,
-        name: str,
-        targets: Sequence[str],
-        pageid: FileId,
-        title: Sequence[n.InlineNode],
-    ) -> None:
-        # If multiple target names are given, prefer placing the one with the most periods
-        # into referring RefRole nodes. This is an odd heuristic, but should work for now.
-        # e.g. if a RefRole links to "-v", we want it to get normalized to "mongod.-v" if that's
-        # what gets resolved.
-        canonical_target_name = max(targets, key=lambda x: x.count("."))
-
-        for target in targets:
-            target = normalize_target(target)
-            key = f"{domain}:{name}:{target}"
-            self.local_definitions[key].append(
-                TargetDatabase.LocalDefinition(canonical_target_name, pageid, title)
-            )
-
-    def reset(self, config: "ProjectConfig") -> None:
-        """Reset this database to a "blank" state with intersphinx inventories defined by
-           the given ProjectConfig instance."""
-        self.intersphinx_inventories.clear()
-        self.local_definitions.clear()
-
-        logger.debug("Loading %s intersphinx inventories", len(config.intersphinx))
-        for url in config.intersphinx:
-            self.intersphinx_inventories[url] = intersphinx.fetch_inventory(url)
-
-    def generate_inventory(self, base_url: str) -> intersphinx.Inventory:
-        targets: Dict[str, intersphinx.TargetDefinition] = {}
-        for key, definitions in self.local_definitions.items():
-            if not definitions:
-                continue
-
-            definition = definitions[0]
-            uri = definition.fileid.without_known_suffix + "/"
-            dispname = "".join(node.get_text() for node in definition.title)
-            domain, role_name, name = key.split(":", 2)
-
-            if not dispname:
-                dispname = name
-
-            base_uri = uri
-            if (domain, role_name) != ("std", "doc"):
-                base_uri += "#$"
-                uri = uri + "#" + make_id(name)
-
-            targets[key] = intersphinx.TargetDefinition(
-                definition.canonical_name,
-                (domain, role_name),
-                -1,
-                base_uri,
-                uri,
-                dispname,
-            )
-        return intersphinx.Inventory(base_url, targets)
-
-    @classmethod
-    def load(cls, config: "ProjectConfig") -> "TargetDatabase":
-        """Create a TargetDatabase with the intersphinx inventories specified by the given
-           ProjectConfig."""
-        db = cls()
-        db.reset(config)
-        return db
-
-
-@dataclass
-class Cache:
-    """A versioned cache that associates a (FileId, int) pair with an arbitrary object and
-       an integer version. Whenever the key is re-assigned, the version is incremented."""
-
-    _cache: Dict[Tuple[FileId, int], object] = field(default_factory=dict)
-    _keys_of_each_fileid: DefaultDict[FileId, Set[int]] = field(
-        default_factory=lambda: defaultdict(set)
-    )
-    _versions: DefaultDict[Tuple[FileId, int], int] = field(
-        default_factory=lambda: defaultdict(int)
-    )
-
-    def __setitem__(self, key: Tuple[FileId, int], value: object) -> None:
-        if key in self._cache:
-            self._cache[key] = value
-        else:
-            self._cache[key] = value
-
-        self._versions[key] += 1
-        self._keys_of_each_fileid[key[0]].add(key[1])
-
-    def __delitem__(self, fileid: FileId) -> None:
-        keys = self._keys_of_each_fileid[fileid]
-        del self._keys_of_each_fileid[fileid]
-        for key in keys:
-            del self._cache[(fileid, key)]
-
-    def __getitem__(self, key: Tuple[FileId, int]) -> Optional[object]:
-        return self._cache.get(key, None)
-
-    def get_versions(self, fileid: FileId) -> Iterator[int]:
-        for key, version in self._versions.items():
-            if key[0] == fileid:
-                yield version
-
-
-class ProjectInterface(Protocol):
-    expensive_operation_cache: Cache
-    targets: TargetDatabase
-
-
-@dataclass
-class EmptyProjectInterface:
-    """An empty ProjectInterface implementation for testing."""
-
-    expensive_operation_cache: Cache
-    targets: TargetDatabase
-
-    def __init__(self) -> None:
-        self.expensive_operation_cache = Cache()
-        self.targets = TargetDatabase()
-
-
-class PendingTask:
-    """A thunk which will be executed in the main process after the full tree is
-       constructed. This should primarily be used to execute tasks which may need
-       to mutate state from the main process (e.g. caches or dependency graphs)."""
-
-    def __init__(self, node: n.Node) -> None:
-        self.node = node
-
-    def __call__(
-        self, diagnostics: List[Diagnostic], project: ProjectInterface
-    ) -> None:
-        """Perform an action in the main process once the tree has been built."""
-        pass
-
-
-@dataclass
-class Page:
-    source_path: Path
-    output_filename: str
-    source: str
-    ast: n.Parent[n.Node]
-    static_assets: Set[StaticAsset] = field(default_factory=set)
-    pending_tasks: List[PendingTask] = field(default_factory=list)
-    category: Optional[str] = None
-    query_fields: Dict[str, SerializableType] = field(default_factory=dict)
-
-    @classmethod
-    def create(
-        self,
-        source_path: Path,
-        output_filename: Optional[str],
-        source: str,
-        ast: Optional[n.Parent[n.Node]] = None,
-    ) -> "Page":
-        if output_filename is None:
-            output_filename = source_path.name
-
-        if ast is None:
-            ast = n.Root((0,), [], {})
-
-        return Page(source_path, output_filename, source, ast)
-
-    def fake_full_path(self) -> PurePath:
-        """Return a fictitious path (hopefully) uniquely identifying this output artifact."""
-        if self.category:
-            # Giza wrote out yaml file artifacts under a directory. e.g. steps-foo.yaml becomes
-            # steps/foo.rst
-            return self.source_path.parent.joinpath(
-                PurePath(self.category), self.output_filename
-            )
-        return self.source_path
-
-    def finish(
-        self, diagnostics: List[Diagnostic], project: Optional[ProjectInterface] = None
-    ) -> None:
-        """Finish all pending tasks for this page. This should be run in the main process."""
-        for task in self.pending_tasks:
-            task(
-                diagnostics, project if project is not None else EmptyProjectInterface()
-            )
-
-        self.pending_tasks.clear()
-
-
 @checked
 @dataclass
 class ProjectConfig:
     root: Path
     name: str
+    fail_on_diagnostics: bool = field(default=False)
     default_domain: Optional[str] = field(default=None)
     title: str = field(default="untitled")
     source: str = field(default="source")